requirements.txt
PySide6==6.5.2
opencv-python==4.7.0.72
Pillow==9.5.0
numpy==1.24.3 